backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
playwright>=1.41.0
pydantic[email]>=2.6.0
pydantic-settings>=2.1.0
asyncpg>=0.29.0
python-multipart>=0.0.9
psycopg2-binary>=2.9.9
httpx>=0.26.0
requests>=2.31.0
alembic>=1.13.1
python-dotenv>=1.0.0
